WIP: Resolve itemframe entity duplication Fixes #405.
A race condition can sometimes result in itemframes and pedestals adding multiple copies of the entity they carry. This resolves the race condition by waiting 15 seconds after the block loads before checking if the entity is present.  This commit also converts the ABM which ran every 15 seconds into an LBM which only runs once each time the block is loaded.  Finally if the block carries extra entities, they are removed.